缺陷UGDJ-533修改完成。 缺陷描述：对象属性的面积单位切换为平方千米时，大小显示错误。
缺陷导致原因：在选择下拉框里初始化添加平方千米单位时，对应的单位值初始化错误，主要原因是千米（10000000）计算平方千米时使用的是int型进行存储，int的最大值只能装下8位数，平方千米是14位数，导致溢出。

缺陷修改：在AreaUnit中将，在初始化KILOMETER((long) UnitValue.KILOMETER * UnitValue.KILOMETER)时，将int转换为long进行存储

+++ Core/src/main/java/com/supermap/desktop/enums/AreaUnit.java
@@ -12,7 +12,8 @@ public enum AreaUnit {
 	CENTIMETER(UnitValue.CENTIMETER * UnitValue.CENTIMETER),
 	DECIMETER(UnitValue.DECIMETER * UnitValue.DECIMETER),
 	METER(UnitValue.METER * UnitValue.METER),
-	KILOMETER(UnitValue.KILOMETER * UnitValue.KILOMETER),
+	//改为long型，因为10000000的平方使用int装不下会产生越界错误
+	KILOMETER((long) UnitValue.KILOMETER * UnitValue.KILOMETER),
 	MILE(UnitValue.MILE * UnitValue.MILE),
 	INCH(UnitValue.INCH * UnitValue.INCH),
 	FOOT(UnitValue.FOOT * UnitValue.FOOT),
